+++ crates/e-navigator-signals/src/request.rs
@@ -5,7 +5,7 @@ use crate::{
     TraceCorrelationKind, TracePeerContext,
 };
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 #[non_exhaustive]
 pub enum ProtocolKind {